requirements.txt
altgraph==0.17.2
anyio==3.6.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
art==5.7
asgiref==3.5.2
asttokens==2.1.0
async-generator==1.10
attrs==21.4.0
awscli==1.25.96
backcall==0.2.0
backports.statistics==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
botocore==1.27.95
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.6
coverage==6.5.0
cryptography==38.0.1
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.2
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fastapi==0.85.0
fasteners==0.14.1
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Mail==0.9.1
fonttools==4.38.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==2.6.6
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.5.1
gspread==5.3.2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.14.0
idna==2.8
ImageHash==4.2.1
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterthemes==0.20.0
jwt==1.3.1
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lesscpy==0.15.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
meld==3.20.2
mistune==2.0.4
monotonic==1.5
more-itertools==4.2.0
moviepy==1.0.3
mwparserfromhell==0.6.4
mysql-connector==2.2.9
mysql-connector-python==8.0.30
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.3
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.3.56
outcome==1.1.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.4
pluggy==1.0.0
ply==3.11
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.2
PyDrive==1.3.1
Pygments==2.13.0
PyGObject==3.36.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.1.3
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2022.6
PyWavelets==1.3.0
pywikibot==8.0.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rdflib==6.2.0
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.6.0
scipy==1.9.0
SecretStorage==2.3.1
selenium==4.3.0
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.12
sqlparse==0.4.3
stack-data==0.6.1
starlette==0.20.4
systemd-python==234
terminado==0.17.0
texttable==1.6.2
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.5.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.26.9
usb-creator==0.3.7
uvicorn==0.18.3
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.7.0
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
Wikidata==0.7.0
wsproto==1.1.0
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
